refactor data processing for ease of new broker data parsing
