+++ src/types/broker.ts
@@ -1,0 +1,8 @@
+export type BrokerCode =
+  | "schwab"
+  | "robinhood"
+  | "etrade"
+  | "fidelity"
+  | "tda"
+  | "vanguard"
+  | (string & {});
